% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognosMashup.R
\docType{class}
\name{cognosMashup}
\alias{cognosMashup}
\title{IBM Cognos Mashup Service interface for R}
\format{An \code{\link{R6Class}} generator object}
\value{
Object of \code{\link{R6Class}} with methods for communication with IBM Cognos Mashup Service
}
\description{
IBM Cognos Mashup Service interface for R
}
\section{Public Members}{

  \describe{
    \item{\code{dispatcher_url}}{Stores IBM Cognos Mashup Service dispatcher URL}
    \item{\code{credential_template}}{Stores IBM Cognos Mashup Service authentication credential template file path}
  }
}

\section{Private Members}{

  \describe{
    \item{\code{namespace}}{Stores authentication namespace}
    \item{\code{username}}{Stures authentication username}
    \item{\code{password}}{Stores authentication password}
  }
}

\section{Public Methods}{

  \describe{
    \item{\code{logon()}}{Logon to the IBM Cognos Mashup Service using the provided credentials (\code{namespace, username, password}).}
    \item{\code{logoff()}}{Logoff from  the IBM Cognos Mashup Servce.}
    \item{\code{get_report_data(report_id, dataset_id, col_names)}}{Retrieve report dataset (\code{dataset_id}) for the specified report (\code{report_id}).}
  }
}

\keyword{data}
